+++ R/backend-teradata.R
@@ -63,14 +63,20 @@ sql_translation.Teradata <- function(con) {
       bitwXor       = sql_prefix("BITXOR", 2),
       bitwShiftL    = sql_prefix("SHIFTLEFT", 2),
       bitwShiftR    = sql_prefix("SHIFTRIGHT", 2),
-      as.numeric    = sql_cast("NUMERIC"),
+      as.numeric    = function(x, digits = 9L) {
+                        digits <- vctrs::vec_cast(digits, integer())
+                        sql_expr(CAST(!!x %as% DECIMAL(12L, !!digits)))
+                      },
       as.double     = sql_cast("NUMERIC"),
       as.character  = sql_cast("VARCHAR(MAX)"),
+      as.Date       = function(x) {
+                        build_sql("DATE ",x)
+                      },
       log10         = sql_prefix("LOG"),
       log           = sql_log(),
       cot           = sql_cot(),
-      quantile = sql_quantile("APPROX_PERCENTILE"),
-      median = sql_median("APPROX_PERCENTILE"),
+      quantile      = sql_quantile("APPROX_PERCENTILE"),
+      median        = sql_median("APPROX_PERCENTILE"),
       nchar         = sql_prefix("CHARACTER_LENGTH"),
       ceil          = sql_prefix("CEILING"),
       ceiling       = sql_prefix("CEILING"),
@@ -81,17 +87,73 @@ sql_translation.Teradata <- function(con) {
                         len <- stop - start + 1
                         sql_expr(SUBSTR(!!x, !!start, !!len))
                       },
-      paste         =  function(...) {
-                        cli_abort(
-                          "{.fun paste} is not supported in this SQL variant, try {.fun paste0} instead"
-                        )
+      startsWith    = function(string, pattern) {
+                        build_sql('CAST(CASE WHEN INSTR(',
+                        string, ', ', pattern,
+                        ") = 1 THEN 1 ELSE 0 END AS INTEGER)")
+                      },
+      paste         = sql_paste_infix(" ", "||", function(x) sql_expr(!!x)),
+      paste0        = sql_paste_infix("", "||", function(x) sql_expr(!!x)),
+      row_number    = win_rank_tdata("ROW_NUMBER"),
+      week          = function(x){
+                        sql_expr(WEEKNUMBER_OF_YEAR(!!x, 'iso'))
+                      },
+      quarter       = function(x) {
+                        build_sql('to_char(',x,",'q')")
                       }
     ),
     sql_translator(.parent = base_odbc_agg,
       var           = sql_aggregate("VAR_SAMP", "var"),
+      row_number    = win_rank_tdata("ROW_NUMBER"),
+      weighted.mean = function(x, w, na.rm = T) {
+                        # nocov start
+                        win_over(
+                          sql_expr(SUM((!!x * !!w))/SUM(!!w)),
+                          win_current_group(),
+                          win_current_group(),
+                          win_current_frame()
+                        )
+                        # nocov end
+                      }
     ),
     sql_translator(.parent = base_odbc_win,
-      var           = win_recycled("VAR_SAMP")
+      var           = win_recycled("VAR_SAMP"),
+      row_number    = win_rank_tdata("ROW_NUMBER"),
+      lead          = function(x, n = 1L, default = NA, order_by = NULL) {
+                        win_over(
+                          sql_expr(LEAD(!!x, !!n, !!default)),
+                          win_current_group(),
+                          order_by %||% win_current_group(),
+                          win_current_frame()
+                        )
+                      },
+      lag           = function(x, n = 1L, default = NA, order_by = NULL) {
+                        win_over(
+                          sql_expr(LAG(!!x, !!as.integer(n), !!default)),
+                          win_current_group(),
+                          order_by %||%  win_current_group(),
+                          win_current_frame()
+                        )
+                      },
+      cumsum        = function(x, order_by = NULL, frame = c(-Inf, 0)) {
+                        win_over(
+                          sql_expr(SUM(!!x)),
+                          win_current_group(),
+                          order_by %||% win_current_group(),
+                          frame %||% win_current_frame()
+                        )
+                      },
+      weighted.mean = function(x, w, na.rm = T) {
+                        # nocov start
+                        win_over(
+                          sql_expr(SUM((!!x * !!w))/SUM(!!w)),
+                          win_current_group(),
+                          win_current_group(),
+                          win_current_frame()
+                        )
+                        # nocov end
+
+                      }
     )
 
   )}
@@ -102,4 +164,21 @@ sql_table_analyze.Teradata <- function(con, table, ...) {
   build_sql("COLLECT STATISTICS ", as.sql(table, con = con) , con = con)
 }
 
-utils::globalVariables(c("ATAN2", "SUBSTR"))
+utils::globalVariables(c("ATAN2", "SUBSTR", "DECIMAL", "WEEKNUMBER_OF_YEAR", "SUM"))
+
+#' @export
+#' @rdname win_over
+win_rank_tdata <- function(f) {
+  force(f)
+  function(order_by = NULL) {
+    order_by <- order_by %||% win_current_group()
+    if (is_empty(order_by)) order_by <- build_sql("(SELECT NULL)")
+
+    win_over(
+      build_sql(dplyr::sql(f), list()),
+      partition = win_current_group(),
+      order = order_by,
+      frame = win_current_frame()
+    )
+  }
+}